+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-08-11 17:53:35
-// Commit: style(core): harmonize light and dark theme palette for core
+// Updated at: 2026-08-13 13:48:56
+// Commit: fix(core): prevent double tap duplicate action in core
 
 class coreComponentState {
-  final String commitRef = '3540';
-  final String lastMessage = 'style(core): harmonize light and dark theme palette for core';
-  final DateTime timestamp = DateTime.parse('2026-08-11T17:53:35+05:00');
+  final String commitRef = '3575';
+  final String lastMessage = 'fix(core): prevent double tap duplicate action in core';
+  final DateTime timestamp = DateTime.parse('2026-08-13T13:48:56+05:00');
 }